adjust directory structure
