+++ source/tests/consistent/common.py
@@ -26,41 +26,18 @@ from dargs import (
     Argument,
 )
 
-from deepmd.common import (
-    make_default_mesh,
-)
-from deepmd.dpmodel.utils.nlist import (
-    build_neighbor_list,
-    extend_coord_with_ghosts,
-)
+INSTALLED_TF = find_spec("tensorflow") is not None
+INSTALLED_PT = find_spec("torch") is not None
 
-if find_spec("tensorflow") is None:
-    if os.environ.get("CI"):
-        raise ImportError("TensorFlow should be tested in the CI")
-    INSTALLED_TF = False
-else:
-    INSTALLED_TF = True
-try:
-    import torch
-except ImportError as e:
-    if os.environ.get("CI"):
-        raise ImportError("PyTorch should be tested in the CI") from e
-    INSTALLED_PT = False
-else:
-    INSTALLED_PT = True
+if os.environ.get("CI") and not (INSTALLED_TF and INSTALLED_PT):
+    raise ImportError("TensorFlow or PyTorch should be tested in the CI")
 
-if INSTALLED_PT:
-    from deepmd.pt.utils.env import DEVICE as PT_DEVICE
-    from deepmd.pt.utils.nlist import build_neighbor_list as build_neighbor_list_pt
-    from deepmd.pt.utils.nlist import (
-        extend_coord_with_ghosts as extend_coord_with_ghosts_pt,
-    )
+
 if INSTALLED_TF:
     from deepmd.tf.common import (
         clear_session,
     )
     from deepmd.tf.env import (
-        GLOBAL_TF_FLOAT_PRECISION,
         default_tf_session_config,
         tf,
     )
@@ -71,7 +48,6 @@ if INSTALLED_TF:
 
 __all__ = [
     "CommonTest",
-    "DescriptorTest",
     "INSTALLED_TF",
     "INSTALLED_PT",
 ]
@@ -338,67 +314,3 @@ class CommonTest(ABC):
         """Clear the TF session."""
         if not self.skip_tf:
             clear_session()
-
-
-class DescriptorTest:
-    """Useful utilities for descriptor tests."""
-
-    def build_tf_descriptor(self, obj, natoms, coords, atype, box, suffix):
-        t_coord = tf.placeholder(GLOBAL_TF_FLOAT_PRECISION, [None], name="i_coord")
-        t_type = tf.placeholder(tf.int32, [None], name="i_type")
-        t_natoms = tf.placeholder(tf.int32, natoms.shape, name="i_natoms")
-        t_box = tf.placeholder(GLOBAL_TF_FLOAT_PRECISION, [9], name="i_box")
-        t_mesh = tf.placeholder(tf.int32, [None], name="i_mesh")
-        t_des = obj.build(
-            t_coord,
-            t_type,
-            t_natoms,
-            t_box,
-            t_mesh,
-            {},
-            suffix=suffix,
-        )
-        return [t_des], {
-            t_coord: coords,
-            t_type: atype,
-            t_natoms: natoms,
-            t_box: box,
-            t_mesh: make_default_mesh(True, False),
-        }
-
-    def eval_dp_descriptor(self, dp_obj: Any, natoms, coords, atype, box) -> Any:
-        ext_coords, ext_atype, mapping = extend_coord_with_ghosts(
-            coords.reshape(1, -1, 3),
-            atype.reshape(1, -1),
-            box.reshape(1, 3, 3),
-            dp_obj.get_rcut(),
-        )
-        nlist = build_neighbor_list(
-            ext_coords,
-            ext_atype,
-            natoms[0],
-            dp_obj.get_rcut(),
-            dp_obj.get_sel(),
-            distinguish_types=True,
-        )
-        return dp_obj(ext_coords, ext_atype, nlist=nlist)
-
-    def eval_pt_descriptor(self, pt_obj: Any, natoms, coords, atype, box) -> Any:
-        ext_coords, ext_atype, mapping = extend_coord_with_ghosts_pt(
-            torch.from_numpy(coords).to(PT_DEVICE).reshape(1, -1, 3),
-            torch.from_numpy(atype).to(PT_DEVICE).reshape(1, -1),
-            torch.from_numpy(box).to(PT_DEVICE).reshape(1, 3, 3),
-            pt_obj.get_rcut(),
-        )
-        nlist = build_neighbor_list_pt(
-            ext_coords,
-            ext_atype,
-            natoms[0],
-            pt_obj.get_rcut(),
-            pt_obj.get_sel(),
-            distinguish_types=True,
-        )
-        return [
-            x.detach().cpu().numpy() if torch.is_tensor(x) else x
-            for x in pt_obj(ext_coords, ext_atype, nlist=nlist)
-        ]
